backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.11
alembic>=1.13.1
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
httpx>=0.26.0
supabase>=2.3.0
python-dotenv>=1.0.0
ollama>=0.1.6
sentence-transformers>=2.3.1
torch>=2.2.0
Pillow>=10.2.0
